refactor: abstract out selector bar to own file
