+++ src/components/context-selection/index.js
@@ -1,3 +1,1 @@
-export { GroupSelector } from './GroupSelector.js'
-export { PeriodSelector } from './PeriodSelector.js'
-export { OrgUnitSelector } from './OrgUnitSelector.js'
+export { ReportParameterSelector } from './ReportParameterSelector.js'
